% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{arithmetic_return}
\alias{arithmetic_return}
\title{One-period arithmetic returns of a given time-series}
\usage{
arithmetic_return(s)
}
\arguments{
\item{s}{the time-series of prices}
}
\value{
vector
}
\description{
{The daily arithmetic returns between two periods is the difference of the prices over the
two periods divided by the first period's price.}
}
