% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{avEndpoint}
\alias{avEndpoint}
\title{Backend function for Alpha-vantage endpoints}
\usage{
avEndpoint()
}
\value{
string of the url "https://www.alphavantage.co/query"
}
\description{
{Returns the endpoint url for Alpha-Vantage's API.}
}
