% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{backtestBeta2Neutral}
\alias{backtestBeta2Neutral}
\title{Backtest a two-asset beta neutral strategy}
\usage{
backtestBeta2Neutral(
  asset,
  rolling,
  bankroll,
  rate,
  numDays,
  sampleSize,
  fraction
)
}
\arguments{
\item{asset}{the high beta asset to long}

\item{rolling}{whether to use a rolling window of size \code{sampleSize}}

\item{bankroll}{initial bankroll}

\item{rate}{the risk-free rate earned on cash}

\item{numDays}{number of days to trade}

\item{sampleSize}{initial training data size/rolling window size}

\item{fraction}{fraction to trade in stocks versus cash}
}
\description{
{Idealized backtest on closing prices under a 2-asset beta neutral strategy.
The estimating of the betas can be done on a rolling basis, optionally. Returns various
measures of performance and plots the total log-growth of the indidivual assets compared to
the benchmark and the beta-neutral portfolio.}
}
