% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantAsset.R
\name{backtestDominantAsset}
\alias{backtestDominantAsset}
\title{Backtest dominant asset strategy}
\usage{
backtestDominantAsset(returns, bankroll = 1, windowSize = 30, limit = 0.5)
}
\arguments{
\item{returns}{arithmetic returns of stocks}

\item{bankroll}{size of initial investment}

\item{windowSize}{rolling window size}

\item{limit}{upper limit of allocation percentage}
}
\value{
xts
}
\description{
{Backtest the dominant asset strategy on close prices.}
}
