% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_backtesting_strategy.R
\name{BACKTESTING_FULL_EQUITY}
\alias{BACKTESTING_FULL_EQUITY}
\title{The default order size, all of the available margin, which is the largest fraction below 1}
\usage{
BACKTESTING_FULL_EQUITY
}
\description{
The default order size, all of the available margin, which is the largest fraction below 1
}
\keyword{internal}
