% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesters.R
\name{backtestPortfolioGBM}
\alias{backtestPortfolioGBM}
\title{Backtester for GBM portfolio strategy}
\usage{
backtestPortfolioGBM(
  stocks,
  rolling = TRUE,
  bankroll = 1500,
  rate = 0,
  restraint = 0.9,
  numDays = 30,
  sampleSize = 30
)
}
\arguments{
\item{stocks}{the data set of stocks}

\item{rolling}{boolean for rolling estimates}

\item{bankroll}{initial bankroll}

\item{rate}{risk-free rate}

\item{restraint}{max percentage of wealth to use}

\item{numDays}{number of days to trade for}

\item{sampleSize}{number of days for initial sample size in training data or rolling window size}
}
\value{
list
}
\description{
{Continuous time idealized backtester on close prices using
optimal allocations recomputed each day under GBM model dynamics.}
}
