% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta2Hedge}
\alias{beta2Hedge}
\title{Compute beta-neutral portfolio for two assets}
\usage{
beta2Hedge(symbol, period = "daily", key = "premium")
}
\arguments{
\item{symbol}{the stock symbol ticker of the market-correlated asset}

\item{period}{the time-series resolution "daily", "intraday", "weekly", or "monthly"}

\item{key}{the tier of Alpha Vantage API key: "premium" or "free"}
}
\value{
vector
}
\description{
{For two assets with one positive beta and one negative beta, the beta neutral portfolio is a simple
linear system that can be solved explicitly. The weights add up to one and the aggregate beta is zero. This function returns
these weights based off historical estimates of the betas.}
}
