% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{betaHedge2}
\alias{betaHedge2}
\title{Compute the beta-neutral portfolio for one market-correlated asset hedged with a inversely correlated asset.}
\usage{
betaHedge2(arithmeticReturns)
}
\arguments{
\item{arithmeticReturns}{the multivariate time-series of returns. Assumes the first column is the benchmark, e.g. S&P 500}
}
\value{
vector of betas
}
\description{
{For two assets with one positive beta and one negative beta, the beta neutral portfolio is a simple
linear system that can be solved explicitly. The weights add up to one and the aggregate beta is zero. This function returns
these weights based off historical estimates of the betas.}
}
