% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Chart a candlestick series of a stock with a TEMA filter and volume}
\usage{
chart(symbol, days_back = 63, tema_window = 21, adj = FALSE)
}
\arguments{
\item{symbol}{the stock symbol}

\item{days_back}{number of days back to view}

\item{tema_window}{the TEMA moving window}

\item{adj}{whether to use adjusted close prices or not (boolean)}
}
\description{
{Chart a candlestick series of a stock price history.
In blue/red is the long/short-term TEMA filter on either the close or adjusted close prices,
and below is the volume.}
}
