% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_company_overview.R
\name{company_overview}
\alias{company_overview}
\title{Get company overview}
\usage{
company_overview(symbol)
}
\arguments{
\item{symbol}{stock ticker to look up}
}
\value{
list
}
\description{
{Return company financial overview plus basic metrics.}
}
