% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{computeAlphas}
\alias{computeAlphas}
\title{Estimate alpha from OLS estimate of a linear regression against benchmark returns}
\usage{
computeAlphas(arithmeticReturns)
}
\arguments{
\item{arithmeticReturns}{the multivariate time-series of returns. Assumes the first column is the benchmark, e.g. S&P 500}
}
\value{
vector
}
\description{
{Compute historical estimates for the alphas of a sample of assets. These are
based on a linear regression of the arithmetic returns of each asset against that of the benchmark index.}
}
