% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_jumps.R
\name{count_jumps}
\alias{count_jumps}
\title{Count the number of jumps per period above a threshold}
\usage{
count_jumps(
  close_prices,
  jump_thresh = 0.01,
  return_type = "arithmetic",
  side = "down",
  period = "weekly"
)
}
\arguments{
\item{close_prices}{the time-series of close prices}

\item{jump_thresh}{the jump-threshold (positive)}

\item{return_type}{the return type "log" or "arithmetic}

\item{side}{the side of the jump: "up", "down" or "both"}

\item{period}{the period, "daily", "weekly", "monthly", "quarterly", "yearly"}
}
\value{
time-series of counts per period
}
\description{
{Counts the number of jumps \eqn{X_n >= a} (and the other respective events) for some jump-threshold \eqn{a} for different periods}
}
