% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{daily_returns}
\alias{daily_returns}
\title{Compute daily returns of a stock}
\usage{
daily_returns(s)
}
\arguments{
\item{s}{the time-series of daily prices}
}
\value{
xts of logarithmic and arithmetic returns
}
\description{
{Computes both the daily log return and daily arithmetic return
of a given price time-series.}
}
