% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_yte.R
\name{date_yte}
\alias{date_yte}
\title{Convert dates to trading years}
\usage{
date_yte(expiry)
}
\arguments{
\item{expiry}{the date to convert, as a string/character, in the \code{yyyy-mm-dd} format.}
}
\value{
numeric, the time until the expiry date in trading years
}
\description{
{Convert time until expiry date into trading years until expiration}
}
