% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketCapPortfolio.R
\name{diverseMarketPortfolio}
\alias{diverseMarketPortfolio}
\title{Diversity weighted market portfolio}
\usage{
diverseMarketPortfolio(market_port, p = 1)
}
\arguments{
\item{market_port}{object returned from \code{getMarketPortfolio}}

\item{p}{diversity index (a number in the unit interval)}
}
\value{
vector
}
\description{
{See SPT (2002) by E. F.}
}
