% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_market_Cap.R
\name{diversity_weighted_port}
\alias{diversity_weighted_port}
\title{Diversity weighted market portfolio}
\usage{
diversity_weighted_port(market_port, p = 1)
}
\arguments{
\item{market_port}{object returned from \code{market_portfolio}}

\item{p}{diversity index (a number in the unit interval)}
}
\value{
vector
}
\description{
{See SPT (2002) by E. F.}
}
