% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtfm.R
\name{dtfm}
\alias{dtfm}
\title{Fit a discrete time daily returns distribution and get log-optimal allocation}
\usage{
dtfm(symbol, distr = "unif", rate = 0)
}
\arguments{
\item{symbol}{stock ticker to look up}

\item{distr}{distribution to fit: "unif", "norm", more to come...}

\item{rate}{risk-neutral rate}
}
\value{
list of fitted parameters, optimal allocation, and optimal growth (annualized)
}
\description{
{Looks up a stock's history on AV, estimates a return distribution
for daily-arithmetic returns, and then computes the optimal log utility criterion.
}
}
