% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtfm.R
\name{dtfm_strategy}
\alias{dtfm_strategy}
\title{Optimal log-growth allocation for single symbol}
\usage{
dtfm_strategy(symbol, rollingWindow = NULL, rate = 0, kf = 1)
}
\arguments{
\item{symbol}{stock to invest}

\item{rollingWindow}{use NULL for entire price history or integer for most recent \code{n} days}

\item{rate}{risk-free bond yield rate}

\item{kf}{reduction for leveraged positions}
}
\value{
vector
}
\description{
{Downloads prices, likelihood ratio tests mixture against stable
distribution then computes Kelly-criterion under results.}
}
