% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverages.R
\name{ema}
\alias{ema}
\title{Compute an exponential moving average of a time-series}
\usage{
ema(x, alpha = 0.5)
}
\arguments{
\item{x}{the time-series of the signal}

\item{alpha}{smoothing parameter, a value in the unit-interval}
}
\value{
vector
}
\description{
{The standard exponential moving average}
}
\details{
{The exponential moving average is defined by \eqn{s_n = \alpha x_n+(1-\alpha)s_{n-1}} for \eqn{n>1}, with some initial value.}
}
