% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtfm.R
\name{fit_dtfm}
\alias{fit_dtfm}
\title{Fit returns distribution to daily-arithmetic returns}
\usage{
fit_dtfm(x, distr = "norm")
}
\arguments{
\item{x}{time series of daily arithmetic returns}

\item{distr}{distribution name: "norm" or "unif"}
}
\value{
vector
}
\description{
Fit returns distribution to daily-arithmetic returns
}
