% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfm.R
\name{fitGBM}
\alias{fitGBM}
\title{Fit parameters of geometric Brownian motion to data}
\usage{
fitGBM(log_returns, time_step = 1/252)
}
\arguments{
\item{log_returns}{the daily log returns of a stock}

\item{time_step}{the time-step between observations, defaults to 1/252.}
}
\value{
vector
}
\description{
{Fit a geometric Brownian motion model to daily log-returns data.}
}
