% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmModel.R
\name{gbm_model}
\alias{gbm_model}
\title{A comprehensive analysis under a GBM model}
\usage{
gbm_model(
  symbol,
  lambda = 0.06,
  days_back = "full",
  period = "daily",
  adj = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{symbol}{stock symbol to analyze}

\item{lambda}{smoothing parameter for EMA estimates in GBM model}

\item{days_back}{number of data points back to look at}

\item{period}{resolution: daily or months}

\item{adj}{whether to use adjusted close prices or not}

\item{envir}{environment to load stock data into}
}
\value{
data.frame
}
\description{
{A in-depth look at a stock under a GBM model and all its associated properties,
like hitting times. The target is to double wealth within a year, and both
the Buy and Hold strategy and the Kelly criterion are compared to this. An update
will come eventually to allow users to choose target wealth goals.}
}
