% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfm.R
\name{gbm_portfolio}
\alias{gbm_portfolio}
\title{Compute optimal log-growth portfolio for multivariate GBM stocks}
\usage{
gbm_portfolio(symbols, rate = 0, long_frac = 1)
}
\arguments{
\item{symbols}{vector of stock symbols}

\item{rate}{riskless rate of return}

\item{long_frac}{fraction of bankroll to use}
}
\value{
list
}
\description{
{Computes both long-only and short-only log-growth optimal
portfolio allocations up to a percentage of total bankroll and the respective
growth rates. }
}
