% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_profile.R
\name{gbm_profile2}
\alias{gbm_profile2}
\title{Analyze a list of stocks under GBM dynamics}
\usage{
gbm_profile2(tickers, rate = 0)
}
\arguments{
\item{tickers}{vector of stock tickers to analyze, at least of length 1}

\item{rate}{risk-neutral rate to benchmark against}
}
\value{
data.frame
}
\description{
{Estimates the parameters for a geometric Brownian motion model
of the stock prices, the Sharpe-ratio, the Kelly-fraction, and growth-rate
with said parameter-estimates. The estimation is done via MLE.}
}
