% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-tw-mops.R
\name{getCategoryCompanyIDs.mops}
\alias{getCategoryCompanyIDs.mops}
\title{getCategoryCompanyIDs.mops}
\usage{
getCategoryCompanyIDs.mops(marketType = "TWSE")
}
\arguments{
\item{marketType}{must be in c("TWSE", "TWOTC", "TWROTC", "TWPUB")}
}
\description{
function which can help you get the company categories and related stock IDs from http://mops.twse.com.tw/mops/web/t51sb01
}
\examples{
getCategoryCompanyIDs.mops("TWOTC")
}
