% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{getCoins}
\alias{getCoins}
\title{Load multiple cryptocurrencies}
\usage{
getCoins(symbols, key = "premium")
}
\arguments{
\item{symbols}{list of cryptocurrency symbols to look up}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
xts
}
\description{
{Returns common sample of historical prices of given portfolio of cryptocurrencies.}
}
