% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-tw-cnyes.R
\name{getCompanyInfo.cnyes}
\alias{getCompanyInfo.cnyes}
\title{getCompanyInfo.cnyes}
\usage{
getCompanyInfo.cnyes(marketType = "sp500")
}
\arguments{
\item{marketType}{must be in c("sp500")}
}
\description{
function which can help you get the information (data.frame) of different companies from http://www.cnyes.com/usastock/hotprice.aspx?page=hot&kind=sp500
}
\examples{
getCompanyInfo.cnyes("sp500")
}
