% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-tw-mops.R
\name{getCompanyInfo.mops}
\alias{getCompanyInfo.mops}
\title{getCompanyInfo.mops}
\usage{
getCompanyInfo.mops(marketType = "TWSE", code = "01")
}
\arguments{
\item{marketType}{must be in c("TWSE", "TWOTC", "TWROTC", "TWPUB")}

\item{code}{codes in http://mops.twse.com.tw/mops/web/t51sb01}
}
\description{
function which can help you get the information (data.frame) of different companies from http://mops.twse.com.tw/mops/web/t51sb01
}
\examples{
getCompanyInfo.mops("TWOTC","02")
}
