% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{getCompanyOverviewAV}
\alias{getCompanyOverviewAV}
\title{Get company overview}
\usage{
getCompanyOverviewAV(symbol, key = "premium")
}
\arguments{
\item{symbol}{stock ticker to look up}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
list
}
\description{
{Return company financial overview plus basic metrics.}
}
