% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kellyWrappers.R
\name{getGBM}
\alias{getGBM}
\title{Estimate parameters of a GBM from historical stock prices}
\usage{
getGBM(symbol = NULL, prices = NULL, period = "daily", key = "premium")
}
\arguments{
\item{symbol}{the stock ticker to look up}

\item{prices}{optinally pass the time-series of prices}

\item{period}{for looking up by symbol, the period between prices (daily, intraday, weekly, monthly)}

\item{key}{premium or free API key for Alpha-Vantage}
}
\value{
vector
}
\description{
{Pass either the stock ticker symbol or the time-series data itself to fit
a Geometric Brownian motion to the data via the basic MLE routine. Returns a vector containing
the mean drift and the volatility; for daily price-series these are annualized, all other periods are not.}
}
