% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketCapPortfolio.R
\name{getMarketCap}
\alias{getMarketCap}
\title{Get market capitalization from AV}
\usage{
getMarketCap(symbol)
}
\arguments{
\item{symbol}{symbol to look up}
}
\value{
numeric
}
\description{
{The share-price multiplied by shares outstanding}
}
