% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-tw-mops.R
\name{getMarketInfo.mops}
\alias{getMarketInfo.mops}
\title{getMarketInfo.mops}
\usage{
getMarketInfo.mops(marketType = "TWSE")
}
\arguments{
\item{marketType}{must be in c("TWSE", "TWOTC", "TWROTC", "TWPUB")}
}
\description{
function which can help you get the information (data.frame) of different market types from http://mops.twse.com.tw/mops/web/t51sb01
}
\examples{
getMarketInfo.mops("TWOTC")
}
