% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketCapPortfolio.R
\name{getMarketPortfolio}
\alias{getMarketPortfolio}
\title{Create market portfolio via market-cap weights}
\usage{
getMarketPortfolio(symbols)
}
\arguments{
\item{symbols}{list of symbols}
}
\value{
vector
}
\description{
{The market portfolio i.e. weights equal to market caps}
}
