% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-tw-mops.R
\name{getMartketTypeCodes.mops}
\alias{getMartketTypeCodes.mops}
\title{getMartketTypeCodes.mops}
\usage{
getMartketTypeCodes.mops()
}
\description{
function which can help you get the codes of different market types in http://mops.twse.com.tw/mops/web/t51sb01
}
\examples{
getMartketTypeCodes.mops()
}
