% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{getStocks}
\alias{getStocks}
\title{Load multiple stocks}
\usage{
getStocks(symbols, key = "premium")
}
\arguments{
\item{symbols}{list of stock symbols to look up}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
xts
}
\description{
{Returns common sample of historical prices of given portfolio.}
}
