% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooOptionChain.R
\name{getYahooOptionChain}
\alias{getYahooOptionChain}
\title{Download option-chain from Yahoo-Finance}
\usage{
getYahooOptionChain(symbol, option_type = "calls", plotSurface = FALSE)
}
\arguments{
\item{symbol}{stock-ticker symbol to look up}

\item{option_type}{"puts" or "calls"}

\item{plotSurface}{boolean for plotting the 3D raw price-surface (Deprecated)}
}
\value{
data.frame
}
\description{
{Download option chains for all available expiries of a given stock ticker,
then merge them by common strike prices into one matrix/data.frame. This is essentially
a wrapper to \code{quantmod::getOptionChain} with some processing/formatting
added afterwards.}
}
