% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kellyWrappers.R
\name{kellyCriterion}
\alias{kellyCriterion}
\title{Compute univariate Kelly-criterion for a given distribution of asset returns.}
\usage{
kellyCriterion(
  model = "gbm",
  rate = 0,
  symbol = NULL,
  prices = NULL,
  period = "daily",
  key = "premium"
)
}
\arguments{
\item{model}{the distribution used for the returns under which the Kelly criterion is to be computed}

\item{rate}{the risk-free rate earned on a money-market account}

\item{symbol}{the stock ticker to look up}

\item{prices}{optinally pass the time-series of prices}

\item{period}{for looking up by symbol, the period between prices (daily, intraday, weekly, monthly)}

\item{key}{premium or free API key for Alpha-Vantage}
}
\value{
numeric
}
\description{
{Compute the Kelly-criterion log-optimal fraction under a variety of models for the distribution
of period asset returns. Models include continuous time GBM, jump-GBM with normally, kou, and uniformly distributed
jumps, as well as discrete-time models such as uniform, normal, gmm, and stable. Pass either
the stock symbol ticker to look up data or pass the price-series itself.}
}
