% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stock.R
\name{load_stock}
\alias{load_stock}
\title{Load a single stock}
\usage{
load_stock(ticker)
}
\arguments{
\item{ticker}{the stock ticker to load}
}
\value{
see quantmod \code{getSymbols}
}
\description{
Load a single stock
}
