% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stock.R
\name{load_stocks}
\alias{load_stocks}
\title{Load multiple stocks}
\usage{
load_stocks(tickers, call_rate = 60)
}
\arguments{
\item{tickers}{the stock tickers to load}

\item{call_rate}{how long to wait between groups of API calls to Alpha Vantage}
}
\value{
see quantmod \code{getSymbols}
}
\description{
Load multiple stocks
}
