% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{log_return}
\alias{log_return}
\title{Daily log-returns}
\usage{
log_return(prices)
}
\arguments{
\item{prices}{the time-series of (adjusted close) prices}
}
\value{
numeric
}
\description{
{Returns the daily log returns of a price series}
}
