% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_market_Cap.R
\name{market_cap}
\alias{market_cap}
\title{Get market capitalization from AV}
\usage{
market_cap(symbol)
}
\arguments{
\item{symbol}{symbol to look up}
}
\value{
numeric
}
\description{
{The share-price multiplied by shares outstanding}
}
