% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_market_Cap.R
\name{market_portfolio}
\alias{market_portfolio}
\title{Create market portfolio via market-cap weights}
\usage{
market_portfolio(symbols)
}
\arguments{
\item{symbols}{list of symbols}
}
\value{
vector
}
\description{
{The market portfolio i.e. weights equal to market caps}
}
