% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertonVsMixture.R
\name{mertonMixtureDecision}
\alias{mertonMixtureDecision}
\title{Merton vs Mixture likelihood ratio test}
\usage{
mertonMixtureDecision(
  logReturns,
  fits,
  spot,
  rate = 0,
  symbol = "",
  plotDensities = TRUE
)
}
\arguments{
\item{logReturns}{time-series of daily log returns}

\item{fits}{named list of \code{merton} parameters (vector) and \code{mixture} parameters (matrix)}

\item{spot}{the current price}

\item{rate}{the discounting rate}

\item{symbol}{relevant stock symbol name}

\item{plotDensities}{boolean for plotting}
}
\value{
list
}
\description{
{Run a likelihood ratio test to reject one of the Merton or mixture model against
the empirical density.}
}
