% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentumRanking.R
\name{monthlyMomentumRankings}
\alias{monthlyMomentumRankings}
\title{Rank a set of stocks via monthly momentum}
\usage{
monthlyMomentumRankings(monthsBack = 6, stocks = NULL, symbols = NULL)
}
\arguments{
\item{monthsBack}{months back to estimate averages of arithmetic returns}

\item{stocks}{data-set of stocks, can be NULL to look up via tickers}

\item{symbols}{tickers of stocks}
}
\value{
data.frame
}
\description{
{Computes monthly returns, averages them over the last few periods, and
sorts them.}
}
