% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logopt.R
\name{optimalMertonTest}
\alias{optimalMertonTest}
\title{Log optimal allocations under Merton jump diffusion model}
\usage{
optimalMertonTest(
  ticker,
  jt = 0.01,
  rate = 0,
  data_back = 180,
  n_tema = 63,
  N = 100
)
}
\arguments{
\item{ticker}{stock ticker to download}

\item{jt}{jump threshold}

\item{rate}{bond rate}

\item{data_back}{data to go back for TEMA plot}

\item{n_tema}{TEMA window}

\item{N}{number of divisions in entropy grid}
}
\value{
list
}
\description{
{Implementation of optimal log utility for Merton jump diffusion}
}
