% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{optimalPortfolio}
\alias{optimalPortfolio}
\title{Optimal long-short portfolios under geometric Brownian motion}
\usage{
optimalPortfolio(
  symbols,
  rate = 0,
  restraint = 1,
  rollingPeriod = 60,
  stocks = NULL
)
}
\arguments{
\item{symbols}{portfolio of stocks}

\item{rate}{risk-free rate on cash}

\item{restraint}{percentage of wealth to go up to}

\item{rollingPeriod}{rolling period size, can be \code{NULL} for full}

\item{stocks}{default NULL, data returned from \code{getStocks()}.}
}
\value{
data.frame
}
\description{
{Kelly-criterion for long and short portfolios under GBM with rolling
estimations. Price data will either be looked up or can be passed.}
}
