% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMAbacktest.R
\name{optimalTEMA}
\alias{optimalTEMA}
\title{Compute the optimal TEMA-window period}
\usage{
optimalTEMA(prices, kelly, rate, bankroll)
}
\arguments{
\item{prices}{time-series of prices}

\item{kelly}{the fixed fraction for reallocation}

\item{rate}{the rate earned on the cash account}

\item{bankroll}{the initial bankroll to invest}
}
\value{
list or data.frame
}
\description{
{Numerically maximize log-growth of the backtest
over the range of TEMA-window sizes for the given fraction to invest.}
}
