% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_DAY_INTERVAL}
\alias{PERFORMANCE_MEASURES_DAY_INTERVAL}
\title{The interval name of daily candles}
\usage{
PERFORMANCE_MEASURES_DAY_INTERVAL
}
\description{
The interval name of daily candles
}
\keyword{internal}
