% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_HISTORICAL_METHOD}
\alias{PERFORMANCE_MEASURES_HISTORICAL_METHOD}
\title{The value at risk method that reads the loss from past returns}
\usage{
PERFORMANCE_MEASURES_HISTORICAL_METHOD
}
\description{
The value at risk method that reads the loss from past returns
}
\keyword{internal}
