% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_MINUTE_INTERVAL_PATTERN}
\alias{PERFORMANCE_MEASURES_MINUTE_INTERVAL_PATTERN}
\title{The pattern a minute interval such as \verb{5minute} matches}
\usage{
PERFORMANCE_MEASURES_MINUTE_INTERVAL_PATTERN
}
\description{
The pattern a minute interval such as \verb{5minute} matches
}
\keyword{internal}
