% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_PARAMETRIC_METHOD}
\alias{PERFORMANCE_MEASURES_PARAMETRIC_METHOD}
\title{The value at risk method that assumes normally distributed returns}
\usage{
PERFORMANCE_MEASURES_PARAMETRIC_METHOD
}
\description{
The value at risk method that assumes normally distributed returns
}
\keyword{internal}
