% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_TRADING_DAYS_PER_YEAR}
\alias{PERFORMANCE_MEASURES_TRADING_DAYS_PER_YEAR}
\title{The number of trading days in a year, used to annualise \code{day} candles}
\usage{
PERFORMANCE_MEASURES_TRADING_DAYS_PER_YEAR
}
\description{
The number of trading days in a year, used to annualise \code{day} candles
}
\keyword{internal}
