% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_performance_measures.R
\name{PERFORMANCE_MEASURES_TRADING_MINUTES_PER_DAY}
\alias{PERFORMANCE_MEASURES_TRADING_MINUTES_PER_DAY}
\title{The number of minutes in one NSE session, from 09:15 to 15:30}
\usage{
PERFORMANCE_MEASURES_TRADING_MINUTES_PER_DAY
}
\description{
The number of minutes in one NSE session, from 09:15 to 15:30
}
\keyword{internal}
