% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverages.R
\name{plotTEMA}
\alias{plotTEMA}
\title{Plot wrapper for TEMAs}
\usage{
plotTEMA(s, data_back = 30, n_tema = 30, symbol)
}
\arguments{
\item{s}{adjusted close price time-series}

\item{data_back}{number of days to look back}

\item{n_tema}{number of days for the approximate window in the tema}

\item{symbol}{symbol name}
}
\value{
null
}
\description{
{Plot wrapper for TEMAs}
}
