% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonJumps.R
\name{pois_jumps}
\alias{pois_jumps}
\title{Ad-hoc Poisson jump model}
\usage{
pois_jumps(symbol, jump_thresh = 0.03, side = "up")
}
\arguments{
\item{symbol}{ticker symbol}

\item{jump_thresh}{threshold of jump size}

\item{side}{direction of jump: up or down}
}
\value{
numeric
}
\description{
Ad-hoc Poisson jump model
}
