% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonJumps.R
\name{poissonJumps}
\alias{poissonJumps}
\title{Ad-hoc Poisson jump model}
\usage{
poissonJumps(symbol, jump_thresh = 0.03, side = "up", envir = parent.frame())
}
\arguments{
\item{symbol}{ticker symbol; can be stock ticker or a cryptocurrency}

\item{jump_thresh}{threshold of jump size}

\item{side}{direction of jump: up or down}

\item{envir}{environment to load stock data into}
}
\value{
numeric
}
\description{
{An ad-hoc model fitting a Poisson distribution to the number of
jumps excess a threshold per week.}
}
\details{
{ A chi-square test is run against the expected Poisson counts of jumps over the
observed number of events. Log-returns are used so that this can be used in an ad-hoc way to assess
fits for a Merton jump diffusion.}
}
