% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantAsset.R
\name{rollingDominantAsset}
\alias{rollingDominantAsset}
\title{Dominant asset via optimal log growth}
\usage{
rollingDominantAsset(returns, windowSize = 30, limit = 0.5)
}
\arguments{
\item{returns}{arithmetic returns of the stocks}

\item{windowSize}{how many days to use in the rolling window}

\item{limit}{the upper limit in percentage of wealth to invest}
}
\value{
list
}
\description{
{Compute the dominant asset via Kelly-criterion.}
}
