% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_av.R
\name{setup_av}
\alias{setup_av}
\title{Setup defaults for quantmod to pull data from Alpha-Vantage}
\usage{
setup_av()
}
\value{
null
}
\description{
{Set defaults for quantmod functions for pulling data from Alpha Vantage.}
}
\details{
{Make sure to register with Alpha Vantage and obtain a free API key to use their data
services. See documentation for \code{getSymbols} or \code{getSymbols.av} for more info, from the
\code{quantmod} package. Then be sure to store this API key as an environment variable, with the name
\code{api.key}.}
}
