% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverages.R
\name{sma}
\alias{sma}
\title{Compute a simple moving average}
\usage{
sma(x, n = 30)
}
\arguments{
\item{x}{the time-series of observations}

\item{n}{the length of the moving window to average over (positive integer)}
}
\value{
vector
}
\description{
{An implementation of the basic simple moving average. A sample mean is computed over a moving window
of a given time series.}
}
\details{
{An error is returned if the length of the moving window is larger than the sample size of the data.}
}
