% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{stock_returns}
\alias{stock_returns}
\title{Daily returns of a set of stocks}
\usage{
stock_returns(stocks, type = "log")
}
\arguments{
\item{stocks}{xts of stocks}

\item{type}{log or arithmetic}
}
\value{
xts
}
\description{
{Daily returns of multiple stocks.}
}
