% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{stockReturns}
\alias{stockReturns}
\title{Daily returns of a set of stocks}
\usage{
stockReturns(stocks, type = "log")
}
\arguments{
\item{stocks}{xts of stocks}

\item{type}{"log" or "arithmetic"}
}
\value{
xts of daily-returns of multiple stocks
}
\description{
{Daily returns of multiple stocks.}
}
