% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_analysis_strategy_backtests.R
\name{StrategyBacktests}
\alias{StrategyBacktests}
\title{Backtests of trading strategies over an instrument's candles}
\description{
A link in the chain of analysis classes, inherited in the end by \code{Instrument}, which supplies \code{prices()}. Its one method, \code{run_backtest()}, runs a strategy written as a subclass of \code{BacktestStrategy} through the package's own backtesting engine, \code{Backtest}, which follows the rules of the Python \code{backtesting} package (version 0.6.5) that the Python library uses.
}
\examples{
\dontrun{
SmaCross <- R6::R6Class(
  "SmaCross",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      self$indicator("fast", talib::SMA(self$data$Close, timePeriod = 10))
      self$indicator("slow", talib::SMA(self$data$Close, timePeriod = 30))
    },
    next_candle = function() {
      if (self$crossover(self$indicators$fast, self$indicators$slow)) {
        self$buy()
      }
    }
  )
)
infosys <- Instrument$new(exchange = "nse", segment = "equities", symbol = "INFY")
statistics <- infosys$run_backtest(
  SmaCross,
  cash = 100000,
  days = 730,
  plot_filename = "sma_cross.html"
)
}

## ------------------------------------------------
## Method `StrategyBacktests$run_backtest()`
## ------------------------------------------------

\dontrun{
MovingAverageCross <- R6::R6Class(
  "MovingAverageCross",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      self$indicator(
        "fast_average",
        talib::SMA(self$data$Close, timePeriod = 10)
      )
      self$indicator(
        "slow_average",
        talib::SMA(self$data$Close, timePeriod = 30)
      )
    },
    next_candle = function() {
      fast <- self$indicators$fast_average
      slow <- self$indicators$slow_average
      if (self$crossover(fast, slow)) {
        self$position$close()
        self$buy()
      } else if (self$crossover(slow, fast)) {
        self$position$close()
        self$sell()
      }
    }
  )
)
infosys <- Equity$new(exchange = "nse", symbol = "INFY")
statistics <- infosys$run_backtest(
  MovingAverageCross,
  cash = 100000,
  days = 730
)
return_percent <- statistics[["Return [\%]"]]
trade_count <- statistics[["# Trades"]]
cat(sprintf("Return: \%.2f\%\% from \%d trades\n", return_percent, trade_count))

BuyAndHold <- R6::R6Class(
  "BuyAndHold",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      invisible(NULL)
    },
    next_candle = function() {
      if (self$position$size == 0) {
        self$buy()
      }
    }
  )
)
nifty <- EquityIndex$new(exchange = "nse", symbol = "NIFTY")
commissions <- c(
  0.0,
  0.001
)
for (commission in commissions) {
  statistics <- nifty$run_backtest(
    BuyAndHold,
    cash = 10000000,
    commission = commission,
    days = 365
  )
  final_equity <- statistics[["Equity Final [$]"]]
  cat(
    sprintf(
      "Commission \%s: ends at \%s\n",
      commission,
      format(round(final_equity, 2), big.mark = ",", nsmall = 2)
    )
  )
}

RelativeStrengthReversal <- R6::R6Class(
  "RelativeStrengthReversal",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      self$indicator(
        "strength",
        talib::RSI(self$data$Close, timePeriod = 14)
      )
    },
    next_candle = function() {
      strength <- tail(self$indicators$strength, 1)
      if (strength < 30 && self$position$size == 0) {
        self$buy()
      } else if (strength > 70 && self$position$size != 0) {
        self$position$close()
      }
    }
  )
)
plot_path <- file.path(tempdir(), "tcs_rsi_backtest.html")
tcs <- Equity$new(exchange = "nse", symbol = "TCS")
statistics <- tcs$run_backtest(
  RelativeStrengthReversal,
  cash = 100000,
  from_date = "2025-01-01",
  to_date = "2025-12-31",
  plot_filename = plot_path
)
measures <- c(
  "Return [\%]",
  "Win Rate [\%]",
  "Max. Drawdown [\%]"
)
str(statistics[measures])
cat(sprintf("Plot written: \%s\n", file.exists(plot_path)))
}
}
\section{Super classes}{
\code{\link[tradeR:PriceAnalysis]{PriceAnalysis}} -> \code{\link[tradeR:PriceStatistics]{PriceStatistics}} -> \code{\link[tradeR:OverlapStudies]{OverlapStudies}} -> \code{\link[tradeR:MomentumIndicators]{MomentumIndicators}} -> \code{\link[tradeR:VolumeIndicators]{VolumeIndicators}} -> \code{\link[tradeR:CycleIndicators]{CycleIndicators}} -> \code{\link[tradeR:PriceTransforms]{PriceTransforms}} -> \code{\link[tradeR:VolatilityIndicators]{VolatilityIndicators}} -> \code{\link[tradeR:StatisticFunctions]{StatisticFunctions}} -> \code{\link[tradeR:MathTransforms]{MathTransforms}} -> \code{\link[tradeR:MathOperators]{MathOperators}} -> \code{\link[tradeR:CandlestickPatterns]{CandlestickPatterns}} -> \code{\link[tradeR:Signals]{Signals}} -> \code{StrategyBacktests}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-StrategyBacktests-run_backtest}{\code{StrategyBacktests$run_backtest()}}
    \item \href{#method-StrategyBacktests-clone}{\code{StrategyBacktests$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceAnalysis" data-id="prices"><a href='../../tradeR/html/PriceAnalysis.html#method-PriceAnalysis-prices'><code>PriceAnalysis$prices()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_high"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_high'><code>PriceStatistics$price_high()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_histogram"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_histogram'><code>PriceStatistics$price_histogram()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_kurtosis"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_kurtosis'><code>PriceStatistics$price_kurtosis()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_low"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_low'><code>PriceStatistics$price_low()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_mean"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_mean'><code>PriceStatistics$price_mean()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_mean_absolute_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_mean_absolute_deviation'><code>PriceStatistics$price_mean_absolute_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_median"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_median'><code>PriceStatistics$price_median()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_quantile"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_quantile'><code>PriceStatistics$price_quantile()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_skewness"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_skewness'><code>PriceStatistics$price_skewness()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_standard_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_standard_deviation'><code>PriceStatistics$price_standard_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_summary"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_summary'><code>PriceStatistics$price_summary()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="price_variance"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-price_variance'><code>PriceStatistics$price_variance()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns'><code>PriceStatistics$returns()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_high"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_high'><code>PriceStatistics$returns_high()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_histogram"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_histogram'><code>PriceStatistics$returns_histogram()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_kurtosis"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_kurtosis'><code>PriceStatistics$returns_kurtosis()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_low"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_low'><code>PriceStatistics$returns_low()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_mean"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_mean'><code>PriceStatistics$returns_mean()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_mean_absolute_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_mean_absolute_deviation'><code>PriceStatistics$returns_mean_absolute_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_median"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_median'><code>PriceStatistics$returns_median()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_quantile"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_quantile'><code>PriceStatistics$returns_quantile()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_skewness"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_skewness'><code>PriceStatistics$returns_skewness()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_standard_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_standard_deviation'><code>PriceStatistics$returns_standard_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_summary"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_summary'><code>PriceStatistics$returns_summary()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="returns_variance"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-returns_variance'><code>PriceStatistics$returns_variance()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_high"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_high'><code>PriceStatistics$volume_high()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_histogram"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_histogram'><code>PriceStatistics$volume_histogram()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_kurtosis"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_kurtosis'><code>PriceStatistics$volume_kurtosis()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_low"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_low'><code>PriceStatistics$volume_low()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_mean"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_mean'><code>PriceStatistics$volume_mean()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_mean_absolute_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_mean_absolute_deviation'><code>PriceStatistics$volume_mean_absolute_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_median"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_median'><code>PriceStatistics$volume_median()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_quantile"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_quantile'><code>PriceStatistics$volume_quantile()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_skewness"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_skewness'><code>PriceStatistics$volume_skewness()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_standard_deviation"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_standard_deviation'><code>PriceStatistics$volume_standard_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_summary"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_summary'><code>PriceStatistics$volume_summary()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_total"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_total'><code>PriceStatistics$volume_total()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volume_variance"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volume_variance'><code>PriceStatistics$volume_variance()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceStatistics" data-id="volumes"><a href='../../tradeR/html/PriceStatistics.html#method-PriceStatistics-volumes'><code>PriceStatistics$volumes()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="bollinger_bands"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-bollinger_bands'><code>OverlapStudies$bollinger_bands()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="double_exponential_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-double_exponential_moving_average'><code>OverlapStudies$double_exponential_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="exponential_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-exponential_moving_average'><code>OverlapStudies$exponential_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="kaufman_adaptive_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-kaufman_adaptive_moving_average'><code>OverlapStudies$kaufman_adaptive_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="mesa_adaptive_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-mesa_adaptive_moving_average'><code>OverlapStudies$mesa_adaptive_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="mid_point"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-mid_point'><code>OverlapStudies$mid_point()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="middle_price"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-middle_price'><code>OverlapStudies$middle_price()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="parabolic_sar"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-parabolic_sar'><code>OverlapStudies$parabolic_sar()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="simple_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-simple_moving_average'><code>OverlapStudies$simple_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="triangular_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-triangular_moving_average'><code>OverlapStudies$triangular_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="triple_exponential_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-triple_exponential_moving_average'><code>OverlapStudies$triple_exponential_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="OverlapStudies" data-id="weighted_moving_average"><a href='../../tradeR/html/OverlapStudies.html#method-OverlapStudies-weighted_moving_average'><code>OverlapStudies$weighted_moving_average()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="absolute_price_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-absolute_price_oscillator'><code>MomentumIndicators$absolute_price_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="aroon"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-aroon'><code>MomentumIndicators$aroon()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="aroon_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-aroon_oscillator'><code>MomentumIndicators$aroon_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="average_directional_movement_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-average_directional_movement_index'><code>MomentumIndicators$average_directional_movement_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="average_directional_movement_index_rating"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-average_directional_movement_index_rating'><code>MomentumIndicators$average_directional_movement_index_rating()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="balance_of_power"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-balance_of_power'><code>MomentumIndicators$balance_of_power()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="chande_momentum_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-chande_momentum_oscillator'><code>MomentumIndicators$chande_momentum_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="commodity_channel_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-commodity_channel_index'><code>MomentumIndicators$commodity_channel_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="directional_movement_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-directional_movement_index'><code>MomentumIndicators$directional_movement_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="minus_directional_indicator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-minus_directional_indicator'><code>MomentumIndicators$minus_directional_indicator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="minus_directional_movement"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-minus_directional_movement'><code>MomentumIndicators$minus_directional_movement()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="momentum"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-momentum'><code>MomentumIndicators$momentum()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="money_flow_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-money_flow_index'><code>MomentumIndicators$money_flow_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="moving_average_convergence_divergence"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-moving_average_convergence_divergence'><code>MomentumIndicators$moving_average_convergence_divergence()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="moving_average_convergence_divergence_extended"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-moving_average_convergence_divergence_extended'><code>MomentumIndicators$moving_average_convergence_divergence_extended()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="percentage_price_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-percentage_price_oscillator'><code>MomentumIndicators$percentage_price_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="plus_directional_indicator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-plus_directional_indicator'><code>MomentumIndicators$plus_directional_indicator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="plus_directional_movement"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-plus_directional_movement'><code>MomentumIndicators$plus_directional_movement()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="rate_of_change"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-rate_of_change'><code>MomentumIndicators$rate_of_change()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="rate_of_change_percent"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-rate_of_change_percent'><code>MomentumIndicators$rate_of_change_percent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="rate_of_change_ratio"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-rate_of_change_ratio'><code>MomentumIndicators$rate_of_change_ratio()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="relative_strength_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-relative_strength_index'><code>MomentumIndicators$relative_strength_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="stochastic_fast_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-stochastic_fast_oscillator'><code>MomentumIndicators$stochastic_fast_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="stochastic_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-stochastic_oscillator'><code>MomentumIndicators$stochastic_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="stochastic_relative_strength_index"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-stochastic_relative_strength_index'><code>MomentumIndicators$stochastic_relative_strength_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="trix"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-trix'><code>MomentumIndicators$trix()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="ultimate_oscillator"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-ultimate_oscillator'><code>MomentumIndicators$ultimate_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MomentumIndicators" data-id="williams_percent_r"><a href='../../tradeR/html/MomentumIndicators.html#method-MomentumIndicators-williams_percent_r'><code>MomentumIndicators$williams_percent_r()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolumeIndicators" data-id="chaikin_accumulation_distribution_line"><a href='../../tradeR/html/VolumeIndicators.html#method-VolumeIndicators-chaikin_accumulation_distribution_line'><code>VolumeIndicators$chaikin_accumulation_distribution_line()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolumeIndicators" data-id="chaikin_accumulation_distribution_oscillator"><a href='../../tradeR/html/VolumeIndicators.html#method-VolumeIndicators-chaikin_accumulation_distribution_oscillator'><code>VolumeIndicators$chaikin_accumulation_distribution_oscillator()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolumeIndicators" data-id="on_balance_volume"><a href='../../tradeR/html/VolumeIndicators.html#method-VolumeIndicators-on_balance_volume'><code>VolumeIndicators$on_balance_volume()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_dominant_cycle_period"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_dominant_cycle_period'><code>CycleIndicators$hilbert_transform_dominant_cycle_period()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_dominant_cycle_phase"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_dominant_cycle_phase'><code>CycleIndicators$hilbert_transform_dominant_cycle_phase()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_phasor_components"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_phasor_components'><code>CycleIndicators$hilbert_transform_phasor_components()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_sine_wave"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_sine_wave'><code>CycleIndicators$hilbert_transform_sine_wave()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_trend_line"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_trend_line'><code>CycleIndicators$hilbert_transform_trend_line()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CycleIndicators" data-id="hilbert_transform_trend_mode"><a href='../../tradeR/html/CycleIndicators.html#method-CycleIndicators-hilbert_transform_trend_mode'><code>CycleIndicators$hilbert_transform_trend_mode()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceTransforms" data-id="average_price"><a href='../../tradeR/html/PriceTransforms.html#method-PriceTransforms-average_price'><code>PriceTransforms$average_price()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceTransforms" data-id="median_price"><a href='../../tradeR/html/PriceTransforms.html#method-PriceTransforms-median_price'><code>PriceTransforms$median_price()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceTransforms" data-id="typical_price"><a href='../../tradeR/html/PriceTransforms.html#method-PriceTransforms-typical_price'><code>PriceTransforms$typical_price()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="PriceTransforms" data-id="weighted_close"><a href='../../tradeR/html/PriceTransforms.html#method-PriceTransforms-weighted_close'><code>PriceTransforms$weighted_close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolatilityIndicators" data-id="average_true_range"><a href='../../tradeR/html/VolatilityIndicators.html#method-VolatilityIndicators-average_true_range'><code>VolatilityIndicators$average_true_range()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolatilityIndicators" data-id="normalized_average_true_range"><a href='../../tradeR/html/VolatilityIndicators.html#method-VolatilityIndicators-normalized_average_true_range'><code>VolatilityIndicators$normalized_average_true_range()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="VolatilityIndicators" data-id="true_range"><a href='../../tradeR/html/VolatilityIndicators.html#method-VolatilityIndicators-true_range'><code>VolatilityIndicators$true_range()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="beta"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-beta'><code>StatisticFunctions$beta()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="correlation_coefficient"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-correlation_coefficient'><code>StatisticFunctions$correlation_coefficient()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="linear_regression"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-linear_regression'><code>StatisticFunctions$linear_regression()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="linear_regression_angle"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-linear_regression_angle'><code>StatisticFunctions$linear_regression_angle()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="linear_regression_intercept"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-linear_regression_intercept'><code>StatisticFunctions$linear_regression_intercept()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="linear_regression_slope"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-linear_regression_slope'><code>StatisticFunctions$linear_regression_slope()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="standard_deviation"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-standard_deviation'><code>StatisticFunctions$standard_deviation()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="StatisticFunctions" data-id="variance"><a href='../../tradeR/html/StatisticFunctions.html#method-StatisticFunctions-variance'><code>StatisticFunctions$variance()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="arc_cosine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-arc_cosine'><code>MathTransforms$arc_cosine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="arc_sine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-arc_sine'><code>MathTransforms$arc_sine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="arc_tangent"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-arc_tangent'><code>MathTransforms$arc_tangent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="ceiling"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-ceiling'><code>MathTransforms$ceiling()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="cosine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-cosine'><code>MathTransforms$cosine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="exponential"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-exponential'><code>MathTransforms$exponential()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="floor"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-floor'><code>MathTransforms$floor()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="hyperbolic_cosine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-hyperbolic_cosine'><code>MathTransforms$hyperbolic_cosine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="hyperbolic_sine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-hyperbolic_sine'><code>MathTransforms$hyperbolic_sine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="hyperbolic_tangent"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-hyperbolic_tangent'><code>MathTransforms$hyperbolic_tangent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="logarithm_base_10"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-logarithm_base_10'><code>MathTransforms$logarithm_base_10()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="natural_logarithm"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-natural_logarithm'><code>MathTransforms$natural_logarithm()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="sine"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-sine'><code>MathTransforms$sine()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="square_root"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-square_root'><code>MathTransforms$square_root()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathTransforms" data-id="tangent"><a href='../../tradeR/html/MathTransforms.html#method-MathTransforms-tangent'><code>MathTransforms$tangent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="add"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-add'><code>MathOperators$add()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="divide"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-divide'><code>MathOperators$divide()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="maximum"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-maximum'><code>MathOperators$maximum()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="maximum_index"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-maximum_index'><code>MathOperators$maximum_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="minimum"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-minimum'><code>MathOperators$minimum()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="minimum_index"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-minimum_index'><code>MathOperators$minimum_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="minimum_maximum"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-minimum_maximum'><code>MathOperators$minimum_maximum()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="minimum_maximum_index"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-minimum_maximum_index'><code>MathOperators$minimum_maximum_index()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="multiply"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-multiply'><code>MathOperators$multiply()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="MathOperators" data-id="subtract"><a href='../../tradeR/html/MathOperators.html#method-MathOperators-subtract'><code>MathOperators$subtract()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_abandoned_baby"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_abandoned_baby'><code>CandlestickPatterns$candle_abandoned_baby()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_advance_block"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_advance_block'><code>CandlestickPatterns$candle_advance_block()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_belt_hold"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_belt_hold'><code>CandlestickPatterns$candle_belt_hold()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_breakaway"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_breakaway'><code>CandlestickPatterns$candle_breakaway()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_closing_marubozu"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_closing_marubozu'><code>CandlestickPatterns$candle_closing_marubozu()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_concealing_baby_swallow"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_concealing_baby_swallow'><code>CandlestickPatterns$candle_concealing_baby_swallow()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_counter_attack"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_counter_attack'><code>CandlestickPatterns$candle_counter_attack()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_dark_cloud_cover"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_dark_cloud_cover'><code>CandlestickPatterns$candle_dark_cloud_cover()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_doji"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_doji'><code>CandlestickPatterns$candle_doji()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_doji_star"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_doji_star'><code>CandlestickPatterns$candle_doji_star()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_dragonfly_doji"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_dragonfly_doji'><code>CandlestickPatterns$candle_dragonfly_doji()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_engulfing"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_engulfing'><code>CandlestickPatterns$candle_engulfing()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_evening_doji_star"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_evening_doji_star'><code>CandlestickPatterns$candle_evening_doji_star()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_evening_star"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_evening_star'><code>CandlestickPatterns$candle_evening_star()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_gravestone_doji"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_gravestone_doji'><code>CandlestickPatterns$candle_gravestone_doji()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_hammer"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_hammer'><code>CandlestickPatterns$candle_hammer()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_hanging_man"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_hanging_man'><code>CandlestickPatterns$candle_hanging_man()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_harami"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_harami'><code>CandlestickPatterns$candle_harami()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_harami_cross"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_harami_cross'><code>CandlestickPatterns$candle_harami_cross()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_high_wave"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_high_wave'><code>CandlestickPatterns$candle_high_wave()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_hikkake"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_hikkake'><code>CandlestickPatterns$candle_hikkake()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_homing_pigeon"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_homing_pigeon'><code>CandlestickPatterns$candle_homing_pigeon()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_identical_three_crows"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_identical_three_crows'><code>CandlestickPatterns$candle_identical_three_crows()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_in_neck"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_in_neck'><code>CandlestickPatterns$candle_in_neck()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_inverted_hammer"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_inverted_hammer'><code>CandlestickPatterns$candle_inverted_hammer()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_kicking"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_kicking'><code>CandlestickPatterns$candle_kicking()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_kicking_by_length"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_kicking_by_length'><code>CandlestickPatterns$candle_kicking_by_length()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_ladder_bottom"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_ladder_bottom'><code>CandlestickPatterns$candle_ladder_bottom()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_long_legged_doji"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_long_legged_doji'><code>CandlestickPatterns$candle_long_legged_doji()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_long_line"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_long_line'><code>CandlestickPatterns$candle_long_line()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_marubozu"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_marubozu'><code>CandlestickPatterns$candle_marubozu()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_mat_hold"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_mat_hold'><code>CandlestickPatterns$candle_mat_hold()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_matching_low"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_matching_low'><code>CandlestickPatterns$candle_matching_low()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_modified_hikkake"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_modified_hikkake'><code>CandlestickPatterns$candle_modified_hikkake()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_morning_star"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_morning_star'><code>CandlestickPatterns$candle_morning_star()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_morning_star_doji"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_morning_star_doji'><code>CandlestickPatterns$candle_morning_star_doji()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_on_neck"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_on_neck'><code>CandlestickPatterns$candle_on_neck()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_piercing"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_piercing'><code>CandlestickPatterns$candle_piercing()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_rickshaw_man"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_rickshaw_man'><code>CandlestickPatterns$candle_rickshaw_man()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_rise_fall_three_methods"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_rise_fall_three_methods'><code>CandlestickPatterns$candle_rise_fall_three_methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_separating_lines"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_separating_lines'><code>CandlestickPatterns$candle_separating_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_shooting_star"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_shooting_star'><code>CandlestickPatterns$candle_shooting_star()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_short_line"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_short_line'><code>CandlestickPatterns$candle_short_line()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_side_by_side_white_lines"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_side_by_side_white_lines'><code>CandlestickPatterns$candle_side_by_side_white_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_spinning_top"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_spinning_top'><code>CandlestickPatterns$candle_spinning_top()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_stalled_pattern"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_stalled_pattern'><code>CandlestickPatterns$candle_stalled_pattern()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_stick_sandwich"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_stick_sandwich'><code>CandlestickPatterns$candle_stick_sandwich()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_takuri"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_takuri'><code>CandlestickPatterns$candle_takuri()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_tasuki_gap"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_tasuki_gap'><code>CandlestickPatterns$candle_tasuki_gap()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_black_crows"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_black_crows'><code>CandlestickPatterns$candle_three_black_crows()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_inside_up_down"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_inside_up_down'><code>CandlestickPatterns$candle_three_inside_up_down()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_line_strike"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_line_strike'><code>CandlestickPatterns$candle_three_line_strike()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_outside_up_down"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_outside_up_down'><code>CandlestickPatterns$candle_three_outside_up_down()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_stars_in_the_south"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_stars_in_the_south'><code>CandlestickPatterns$candle_three_stars_in_the_south()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_three_white_soldiers"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_three_white_soldiers'><code>CandlestickPatterns$candle_three_white_soldiers()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_thrusting_pattern"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_thrusting_pattern'><code>CandlestickPatterns$candle_thrusting_pattern()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_tristar"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_tristar'><code>CandlestickPatterns$candle_tristar()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_two_crows"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_two_crows'><code>CandlestickPatterns$candle_two_crows()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_unique_three_river"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_unique_three_river'><code>CandlestickPatterns$candle_unique_three_river()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_up_side_down_side_gap_three_methods"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_up_side_down_side_gap_three_methods'><code>CandlestickPatterns$candle_up_side_down_side_gap_three_methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="CandlestickPatterns" data-id="candle_up_side_gap_two_crows"><a href='../../tradeR/html/CandlestickPatterns.html#method-CandlestickPatterns-candle_up_side_gap_two_crows'><code>CandlestickPatterns$candle_up_side_gap_two_crows()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="Signals" data-id="is_cross_over"><a href='../../tradeR/html/Signals.html#method-Signals-is_cross_over'><code>Signals$is_cross_over()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="tradeR" data-topic="Signals" data-id="is_cross_under"><a href='../../tradeR/html/Signals.html#method-Signals-is_cross_under'><code>Signals$is_cross_under()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyBacktests-run_backtest"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyBacktests-run_backtest}{}}}
\subsection{\code{StrategyBacktests$run_backtest()}}{
  Runs a strategy over the candles in a range, optionally saving a plot of the result.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StrategyBacktests$run_backtest(
  strategy,
  cash = 10000,
  commission = 0,
  margin = 1,
  trade_on_close = FALSE,
  hedging = FALSE,
  exclusive_orders = FALSE,
  plot_filename = NULL,
  interval = "day",
  from_date = NULL,
  to_date = NULL,
  days = NULL,
  adjusted = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{strategy}}{The R6 class generator of the strategy to run, a subclass of \code{BacktestStrategy}.}
      \item{\code{cash}}{The numeric starting cash.}
      \item{\code{commission}}{The numeric commission charged on each trade, as a fraction of its value.}
      \item{\code{margin}}{The numeric margin required, as a fraction, where 1.0 means no leverage.}
      \item{\code{trade_on_close}}{A logical that is \code{TRUE} to fill market orders at the current candle's close rather than the next candle's open.}
      \item{\code{hedging}}{A logical that is \code{TRUE} to allow long and short trades at the same time.}
      \item{\code{exclusive_orders}}{A logical that is \code{TRUE} to close the open trade whenever a new order is placed.}
      \item{\code{plot_filename}}{The character path of an HTML file to write the plot to, or \code{NULL} to skip the plot. The file is not opened. The plot is a plain HTML page of line charts and tables rather than backtesting.py's interactive Bokeh chart.}
      \item{\code{interval}}{The character candle interval, such as \code{"day"} or \code{"5minute"}.}
      \item{\code{from_date}}{The first day of the range as a \code{Date} or a \code{"YYYY-MM-DD"} character value, or \code{NULL} when \code{days} is given.}
      \item{\code{to_date}}{The last day of the range as a \code{Date} or a \code{"YYYY-MM-DD"} character value, or \code{NULL} when \code{days} is given.}
      \item{\code{days}}{The integer number of days to count back from today, or \code{NULL} when \code{from_date} and \code{to_date} are given.}
      \item{\code{adjusted}}{A logical that is \code{TRUE} for prices adjusted for splits and bonuses.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Details}{
    Errors: signals \code{UnifiedBrokerInterfaceError} when UBI refused the request or could not be reached; \code{TypeError} when \code{strategy} does not inherit \code{BacktestStrategy}; \code{ValueError} when the cash, commission or margin is outside its allowed range.
  }
  \subsection{Returns}{
    A named list of the backtest's statistics, with the same names and order as the pandas Series backtesting.py returns, such as \code{"Return [\%]"}, \code{"Sharpe Ratio"} and \code{"# Trades"}, plus \code{"_strategy"}, \code{"_equity_curve"} and \code{"_trades"}; or \code{NULL} when UBI has no candles for the range.
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{MovingAverageCross <- R6::R6Class(
  "MovingAverageCross",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      self$indicator(
        "fast_average",
        talib::SMA(self$data$Close, timePeriod = 10)
      )
      self$indicator(
        "slow_average",
        talib::SMA(self$data$Close, timePeriod = 30)
      )
    },
    next_candle = function() {
      fast <- self$indicators$fast_average
      slow <- self$indicators$slow_average
      if (self$crossover(fast, slow)) {
        self$position$close()
        self$buy()
      } else if (self$crossover(slow, fast)) {
        self$position$close()
        self$sell()
      }
    }
  )
)
infosys <- Equity$new(exchange = "nse", symbol = "INFY")
statistics <- infosys$run_backtest(
  MovingAverageCross,
  cash = 100000,
  days = 730
)
return_percent <- statistics[["Return [\%]"]]
trade_count <- statistics[["# Trades"]]
cat(sprintf("Return: \%.2f\%\% from \%d trades\n", return_percent, trade_count))

BuyAndHold <- R6::R6Class(
  "BuyAndHold",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      invisible(NULL)
    },
    next_candle = function() {
      if (self$position$size == 0) {
        self$buy()
      }
    }
  )
)
nifty <- EquityIndex$new(exchange = "nse", symbol = "NIFTY")
commissions <- c(
  0.0,
  0.001
)
for (commission in commissions) {
  statistics <- nifty$run_backtest(
    BuyAndHold,
    cash = 10000000,
    commission = commission,
    days = 365
  )
  final_equity <- statistics[["Equity Final [$]"]]
  cat(
    sprintf(
      "Commission \%s: ends at \%s\n",
      commission,
      format(round(final_equity, 2), big.mark = ",", nsmall = 2)
    )
  )
}

RelativeStrengthReversal <- R6::R6Class(
  "RelativeStrengthReversal",
  inherit = BacktestStrategy,
  public = list(
    initialize_strategy = function() {
      self$indicator(
        "strength",
        talib::RSI(self$data$Close, timePeriod = 14)
      )
    },
    next_candle = function() {
      strength <- tail(self$indicators$strength, 1)
      if (strength < 30 && self$position$size == 0) {
        self$buy()
      } else if (strength > 70 && self$position$size != 0) {
        self$position$close()
      }
    }
  )
)
plot_path <- file.path(tempdir(), "tcs_rsi_backtest.html")
tcs <- Equity$new(exchange = "nse", symbol = "TCS")
statistics <- tcs$run_backtest(
  RelativeStrengthReversal,
  cash = 100000,
  from_date = "2025-01-01",
  to_date = "2025-12-31",
  plot_filename = plot_path
)
measures <- c(
  "Return [\%]",
  "Win Rate [\%]",
  "Max. Drawdown [\%]"
)
str(statistics[measures])
cat(sprintf("Plot written: \%s\n", file.exists(plot_path)))
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyBacktests-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyBacktests-clone}{}}}
\subsection{\code{StrategyBacktests$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StrategyBacktests$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
