% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverages.R
\name{tema}
\alias{tema}
\title{Compute a so-called triple exponential moving average of a time-series}
\usage{
tema(x, alpha)
}
\arguments{
\item{x}{the time-series of the signal}

\item{alpha}{smoothing parameter, a value in the unit-interval}
}
\value{
vector
}
\description{
{The triple exponential moving average}
}
\details{
{See wikipedia for triple exponential moving average (not triple exponential smoothing)}
}
