% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMAbacktest.R
\name{temaCrossoverBacktest}
\alias{temaCrossoverBacktest}
\title{Backtest fixed fractional reallocation with TEMA exits}
\usage{
temaCrossoverBacktest(N, prices, kelly, rate, bankroll, plotPaths = FALSE)
}
\arguments{
\item{N}{number of days for the moving window in the TEMA filter}

\item{prices}{the time series of prices}

\item{kelly}{the fraction to invest}

\item{rate}{the rate earned on cash account}

\item{bankroll}{the initial bankroll to invest}

\item{plotPaths}{boolean to plot sample paths}
}
\value{
list
}
\description{
{Backtest fixed fractiona reallocation with TEMA exit signals
on a given time series of prices. Assumes perfect trading at close prices with no
transaction costs, etc.}
}
