% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_yte.R
\name{time_yte}
\alias{time_yte}
\title{Convert dates to trading years}
\usage{
time_yte(expiry)
}
\arguments{
\item{expiry}{the date to convert, as a string/character, in the \code{yyyy-mm-dd} format.}
}
\value{
numeric, the time until the expiry date in trading years
}
\description{
{Convert time until expiry date into trading years until expiration. NOTE: this function includes
the exact time-difference but does not account for trading hours.}
}
