% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_profile.R
\name{gbm_profile}
\alias{gbm_profile}
\title{Analyze a list of stocks under GBM dynamics}
\usage{
gbm_profile(tickers, rate = 0)
}
\arguments{
\item{tickers}{vector of stock tickers to analyze, at least of length 1}

\item{rate}{risk-neutral rate to benchmark against}
}
\value{
data.frame containing \code{drift}, the annual mean drift rate \eqn{\mu},
\code{volat} the annual volatility \eqn{\sigma}, \code{sharpe} the Sharpe-ratio \eqn{\lambda := (\mu-r)/\sigma},
\code{kelly} the Kelly-fraction \eqn{(\mu-r)/\sigma^2}, and \code{entropy} the annual growth rate under
the Kelly-fraction, \eqn{0.5(r+0.5\lambda^2)}.
}
\description{
{Estimates the parameters for a geometric Brownian motion model
of the stock prices, the Sharpe-ratio, the Kelly-fraction, and growth-rate
with said parameter-estimates. The estimation is done via MLE.}
}
\details{
{Currently, each stock ticker's historical data is pulled in full,
so direct comparisons of estimations and statistics is not actually sound, due
to the different time horizons and resolution of the sample.}
}
