% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_time_series.R
\name{load_stocks}
\alias{load_stocks}
\title{Load multiple stocks}
\usage{
load_stocks(symbols)
}
\arguments{
\item{symbols}{list of stock symbols to look up}
}
\value{
xts
}
\description{
{Returns common sample of historical prices of given portfolio.}
}
