% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonJumps.R
\name{poissonJumps}
\alias{poissonJumps}
\title{Ad-hoc Poisson jump model}
\usage{
poissonJumps(symbol, jump_thresh = 0.03, side = "up")
}
\arguments{
\item{symbol}{ticker symbol}

\item{jump_thresh}{threshold of jump size}

\item{side}{direction of jump: up or down}
}
\value{
numeric
}
\description{
{An ad-hoc model fitting a Poisson distribution to the number of
jumps excess a threshold per week.}
}
